wip: deployment

+++ deploy/src/main/kotlin/de/mein/deploy/Deploy.kt
@@ -2,6 +2,9 @@ package de.mein.deploy
 
 import de.mein.Lok
 import de.mein.konsole.Konsole
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withTimeout
 import java.io.File
 import java.util.*
 
@@ -16,24 +19,78 @@ class Deploy(private val deploySettings: DeploySettings) {
         fetch()
         val gradle = "./gradlew"
 
-        runProcesses("test",
-                Processor(gradle, ":sql:test"),
-                Processor(gradle, ":auth:test"),
-                Processor(gradle, ":calendar:test"),
-                Processor(gradle, ":contacts:test"),
-                Processor(gradle, ":drive:test"),
-                Processor(gradle, ":konsole:test"),
-                Processor(gradle, ":miniserver:test"),
-                Processor(gradle, ":serialize:test"),
-                Processor(gradle, ":sql:test"))
+//        runProcesses("test",
+//                Processor(gradle, ":auth:test"),
+//                Processor(gradle, ":calendar:test"),
+//                Processor(gradle, ":contacts:test"),
+//                Processor(gradle, ":drive:test"),
+//                Processor(gradle, ":konsole:test"),
+//                Processor(gradle, ":miniserver:test"),
+//                Processor(gradle, ":serialize:test"),
+//                Processor(gradle, ":sql:test"))
 
-        runProcesses("build", Processor(gradle, ":fxbundle:buildFxJar"),
+        runProcesses("build",
+                Processor(gradle, ":fxbundle:buildFxJar"),
                 Processor(gradle, ":app:assembleDebug"),
                 Processor(gradle, ":miniserver:buildServerJar"))
 
+//        runProcesses("debug", Processor(gradle, ":miniserver:buildServerJar"))
 
         val serverDir = File("deploy${File.separator}server")
-        serverDir.deleteRecursively()
+        Lok.debug("setting up deployed dir ${serverDir.absolutePath}")
+        if (serverDir.exists()) {
+            val stopFile = File(serverDir, "stop.input")
+            if (stopFile.exists() && stopFile.canWrite()) {
+                GlobalScope.launch {
+                    withTimeout(1000) {
+                        stopFile.outputStream().write("stop".toByteArray())
+                        stopFile.outputStream().close()
+                    }
+                }
+            }
+            serverDir.deleteRecursively()
+        }
+        serverDir.mkdirs()
+        val serverFilesDir = File(serverDir, "files")
+        serverFilesDir.mkdirs()
+        val target = serverFilesDir.absolutePath
+        val workPath = workingDir.absolutePath + File.separator
+//        runProcesses("debug copy",
+//                Processor("/bin/sh", "-c", "cp \"${workPath}miniserver/build/libs/\"* \"${serverDir.absolutePath}\""))
+
+        runProcesses("copying",
+                Processor("/bin/sh", "-c", "cp \"${workPath}miniserver/build/libs/\"* \"${serverDir.absolutePath}\""),
+                Processor("/bin/sh", "-c", "cp \"${workPath}fxbundle/build/libs/\"* \"$target\""),
+                Processor("/bin/sh", "-c", "cp \"${workPath}app/build/outputs/apk/debug/\"* \"$target\""),
+                Processor("/bin/sh", "-c", "rm \"$target/output.json\""
+                ))
+
+        // start first jar in serverDir
+        val serverJar = serverDir.listFiles().first()
+        Lok.debug("starting server jar ${serverJar.absolutePath}")
+//        runProcesses("launch",
+//                Processor("java", "-jar", serverJar.absolutePath, "-http", "-pipes","&"))
+        Processor("java", "-jar", serverJar.absolutePath, "-http", "-pipes", "-dir", serverDir.absolutePath, "&", "detach").run(false)
+
+        //"-jar", serverJar.absolutePath, "-http", "-pipes", "&"
+//        Processor("/bin/sh", "-c", "java -jar ${serverJar.absolutePath}").run(wait = true)
+//        arrayOf()
+//        ProcessBuilder(*arrayOf("a","b"))
+//        val process = ProcessBuilder("/bin/sh", "-c", "java -jar ${serverJar.absolutePath}","&")
+//                .redirectOutput(ProcessBuilder.Redirect.PIPE)
+//                .redirectError(ProcessBuilder.Redirect.PIPE).start()
+//        process.waitFor()
+//        val exit = process.exitValue()
+//        process.inputStream.bufferedReader().lines().forEach { println(it) }
+//        if (exit != 0) {
+//            Lok.debug("command failed with $exit")
+//            process.inputStream.bufferedReader().lines().forEach { Lok.error(it) }
+//            process.errorStream.bufferedReader().lines().forEach { Lok.error(it) }
+//            error("failed")
+//        }
+        Lok.debug("command succeeded")
+
+        Lok.debug("done")
     }
 
     private fun runProcesses(name: String, vararg processors: Processor) {
